Simplified mario checks
